requirements.txt
streamlit
pandas
supabase
pypff-python
openai
python-dotenv
